Support for Big SSD
-Tested with up to 2TB disk size
-Added support for Delay mode:
    1.real - backwards compatible, same delay that was
      implemented until now
    2.emulated - no delay, execution of test continues normally,
      delay is only logged.
-All relevant variable in code were updated to the appropriate kind,
 size of variable was increased to accommodate larger disk, and
 signed/unsigned was changed depending on actual use of variable
-Calculation of ssd_stats: utilization and read/write speed was fixed
 to account for multiple rt_analyzers
-Added ssd_stats: occupied_pages, logical_write_count,
 block_erase_count, channel_switch_to_write, channel_switch_to_read
-Added mechanism for syncing to log without arbitrary time sleep
-Updated GC: Victim list is now sorted by valid_page_nb, accelerating
 GC algorithm for large disk sizes.
-Update legacy tests for new changes
-ssd_write_read test now has tests for checking all ssd_stats
 during test of read, write, mixed read/write and GC tests

+++ eVSSIM/tests/host/offline_logger_tests.cc
@@ -221,7 +221,7 @@ void flipAuto();
 
         // In case of r/w mode: read / write num_blocks
 
-        for (int i = 0; i < PAGES_IN_SSD; i++) {
+        for (uint64_t i = 0; i < PAGES_IN_SSD; i++) {
             switch (mode) {
                 case MODE_R:
                     SSD_PAGE_READ(CALC_FLASH(i), CALC_BLOCK(i), i, 0, READ);
